Changes in Beautification of codes Campaign Module

+++ Modules/Campaign/Http/Controllers/CampaignController.php
@@ -5,6 +5,7 @@ namespace Modules\Campaign\Http\Controllers;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Routing\Controller;
 use Illuminate\Http\Request;
+use Modules\Campaign\Entities\Campaign;
 use Modules\Campaign\Http\Requests\StoreCampaignRequest;
 use Modules\Campaign\Http\Services\CampaignService;
 
@@ -28,6 +29,17 @@ class CampaignController extends Controller
     public function index(Request $request): JsonResponse
     {
 
+        $user = auth('sanctum')->user();
+
+        // return error if user is not a brand
+        if ($user->cannot('viewAny', Campaign::class)) {
+            return response()->json([
+                'res' => false,
+                'msg' => 'User is not authorized !',
+                'data' => ""
+            ]);
+        }
+        $request->request->add(['user_id' => $user->id]);
         $response = $this->campaignService->getCampaigns($request);
 
         return response()->json($response);
@@ -41,6 +53,17 @@ class CampaignController extends Controller
      */
     public function store(StoreCampaignRequest $request): JsonResponse
     {
+        $user = auth('sanctum')->user();
+
+        // return error if user cannot create campaign
+        if ($user->cannot('create', Campaign::class)) {
+            return response()->json([
+                'res' => false,
+                'msg' => 'User is not authorized !',
+                'data' => ""
+            ]);
+        }
+        $request->request->add(['user_id' => $user->id]);
         $response = $this->campaignService->store($request->validated());
 
         return response()->json($response);
@@ -55,6 +78,18 @@ class CampaignController extends Controller
     public function show(string $campaignKey): JsonResponse
     {
 
+        $user = auth('sanctum')->user();
+        $campaign = Campaign::where('campaign_key', $campaignKey)->first();
+
+        // return error if user not created the campaign
+        if ($user->cannot('view', $campaign)) {
+            return response()->json([
+                'res' => false,
+                'msg' => 'User is not authorized !',
+                'data' => ""
+            ]);
+        }
+
         $response = $this->campaignService->get($campaignKey);
 
         return response()->json($response);
@@ -68,6 +103,17 @@ class CampaignController extends Controller
      */
     public function destroy(string $campaignKey): JsonResponse
     {
+        $user = auth('sanctum')->user();
+        $campaign = Campaign::where('campaign_key', $campaignKey)->first();
+
+        // return error if user not created the campaign
+        if ($user->cannot('delete', $campaign)) {
+            return response()->json([
+                'res' => false,
+                'msg' => 'User is not authorized !',
+                'data' => ""
+            ]);
+        }
 
         $response = $this->campaignService->delete($campaignKey);
